abstract out logging
